Remove option to automerge JavaScript devDependencies
Updates of devDependencies usually contain changes that should be
reviewed. Just a green build is not enough.

A manual review is especially important to see if a dependency has a new
feature we want to use.

The change is also necessary because we'll soon start to use
[renovate-approve-bot](https://github.com/renovatebot/renovate-approve-bot)
which would otherwise merge all devDependencies updates.