Revert "Revert "renovate: Disable stale pull request rebasing""
This reverts commit 26bd06def028dacdfefd1aa0e65bc1d0d2a99cce.

Rebasing all pull requests of renovate automatically leads to annoying
spam in Slack.

It also wastes tons of energy since CI is run without anybody caring
about it.

It should be turned off.

To avoid the problem described in the revert, I suggest two possible
approaches to a solution:

a) think before you merge and update
b) enable the "Require pull requests to be up-to-date" check