Tighten schedule for CircleCI Docker digest updates
CircleCI Docker images are automatically build every night at 00:00 UTC.
This translates to 01:00 CET or 02:00 CEST, depending on the date.

As a result Renovate creates two pull requests. A first one with the
digest before the nightly build and another after it.

Configure digest schedule to run only between 03:00 and 06:00 CET/CEST.

https://github.com/circleci/circleci-images/blob/master/.circleci/config.yml#L428