Quench ajv warnings related to $ref
Renovate uses `"$ref": "#"` in many places of its schema to indicate
that everything which is permissible as top level config is also
valid in a subtree.

According to https://tools.ietf.org/id/draft-pbryan-zyp-json-ref-03.html#rfc.section.3
(apparently this feature never made it past draft stage but is still
in active use) every keyword except `"$ref"` should be ignored when
this feature is used.

However, Renovate intentionally specifies keywords in addition to "$ref".

While this might not be a standard conformant schema, we can validate it
as intended by Renovate by setting `extendRefs` to `true` and not
ignoring those keywords.

As the Renovate schema is autogenerated, and as the option only takes
more validation options into account, there seems little risk associated
with enabling it.