Deactivate Node.js Docker image updates to "Current" version
Instead of having ongoing pull requests to update the Node.js Docker
image to the "Current" version we should actively decide when we want to
update to the next stable version.

This behavior matches Renovate's default configuration of the
`supportPolicy` option. Renovate will only update to "Current" once it
has been promoted to LTS.

https://docs.renovatebot.com/node/#configuring-support-policy