Add poetry preset
We export our Poetry requirements as `requirements.txt` for deployments
to Heroku.

We don't want Renovate to upgrade the `requirements.txt` file in those
cases, as it should always follow the contents of our Poetry
requirements.

Create a preset that is the standard preset, but with the
`pip_requirements` manager disabled.